Removing (without replacement) slotPlayerKilled(), which displayed the GameOverDialog if only one player was left in the game.
It is completely pointless to port this over to the new widget design, as it
needs to be redone anyway, due to our winning conditions. The check whether the
game is over should happen in the canvas event listener which eventually should
pause the game and emit a signal which then displays the GameOverDialog.
If I would port the current code now, then we would throw 100% of that port away,
once we get a proper implementation and that's not worth it.