fixing editor crash/hang when changing the local player
this fix is a hack. in slotEditorChangeLocalPlayer() we use a static variable
that is set to the new local player and then we call the method again using a
QTimer. the player is the previously set static variable then.
this is necessary, because the big display recreates the menu when changing the
local player and libufo doesn't like it when it's widgets are deleted while the
event loop is being processed (which is the case here, as the slot is called by
a menu item).

this hack is the easy solution to this problem and most probably it will remain
the final one, too. it should not have weird side effects and therefore the hack
is controllable. furthermore the problem occurs at this very special case only,
so changing the design would be overkill.