remove the NoTree variants.
it's very sad that this code is kinda lost now, as there are a lot of parts that
would be interesting for other problems. But I doubt anyone will remember that
that code once existed...
Unfortunately the code could cause problems now, as porting the big display to a
nicer design is more difficult when this code is still there. So it must go.

Of course there is no feature loss at all, as the NoTree variants were totally
unused anyway.