implementing BosonMenuInput.
Logic code for all menu items resides in this KGameIO now. For some items, it
just forwards a signal (or emits a slightly modified signal) and the actual code
is in the big display. This is the case for actions that actually belong into
the big display.

Most menu items are untested now (we have way too many to test them all), so
please let me know if some stop working. Exceptions are edit minerals/oil of a
player (in editor mode) and unfogging - these are known not to work currently.