adding BoVeryFastGroundRenderer.
This is the fastest of all possible ground renderers (excluding the "render
nothing" ground renderer that would theoretically be possble). Instead of the
ground textures it uses the closest color only. Furthermore it disables the FOW
texture and uses a more aggressive LOD.

This renderer should not be used by normal players - it is very fast, but also
looks very bad. It is meant for debugging primarily and even then only when
debugging with software rendering.
In software rendering this renderer makes the difference between "barely
playable" and "unusable". With this ground renderer (and --nomodels) ufo and
water take most of the rendering time.