Adding a "edit signals and slots" dialog. In this dialog you can add signals and slots and additionally normal methods.
The slots/methods will be pure virtual and therefore _must_ be implemented by
derived classes.
This feature is intended to be used to define the interface of the class, so
that direct use of the member variables should be kept at a minimum (as they are
much more likely to be changed).

The dialog also supports connections. Currently the sender/receiver/signal/slot
(if applicable) have to be entered into a free-text lineedit. One day we may
replace this by a combobox where available widgets/signals/slots can be chosen
from.