get rid of borderlayouts in the welcome widgets.
a border layout is pretty easy to use and looks very powerful and promising.
however there are a few traps in it (which do not come from the implementation!)
that are hard to come by and are usually unexpected.
therefore i think it's better to avoid border layouts in general.

additionally the boxlayouts are now so welll implemented that they simply look
much better :-)