Adding the FPS counter NG. It is not yet used in current cvs, it will be once I commit my ufo port patch.
The counter is supposed to be used in GLWidget::slotUpdateGL(), i.e. before and
after paintGL() is called, so that the time spent in swapBuffer() is measured as
well.
Before the original slotUpdateGL() implementation is called, startFrame() should
be called and after slotUpdateGL(), endFrame().

The class stores the start/end times internally (for up to 30 seconds by
default) and is therefore able to give on-the-fly fps values with a very good
precision depending on the requirements (if you want to know the current FPS
value you query for the frames of the previous 1 or 2 seconds, if you want to
know the average FPS, you query the frames of the previous 30 seconds).

Additionally the class supports caching - when cachedFPS() is called, it updates
the FPS only once per second, otherwise returns the previous value.
This is especially meant to provide the GUI with a value that does not change
every seconds (note that the speed difference is minor and should not matter
anyway)

The main FPS counter has a "frontend" class BosonGameFPSCounter, that is
supposed to be used outside of the GL widget. It does provide const (i.e.
readonly) access to the FPS counter only, and provides write access to a few
methods that are relevant in the game widgets only.
This class can in particular be used to provide additional information, such as
"average time spent in canvas renderer" or so.

This class has been designed to provide as many information as possible while
providing stability. The previous class sometimes permanently displayed an FPS
of 0.0, that should not happen with this class anymore.