implementing Rivo's idea about an FPS graph: Every 100ms a "snapshot" of the current FPS is taken and drawn onto the screen in every frame as a line strip.
It looks pretty nice, however a label saying that the displayed graph describes
the FPS is still missing. A few labels giving actual numbers to some of the
points would be nice, too.

Note that although it is named "FPSGraph", it may be used for any other kinds of
data, too.

The update period of 100ms seems to be pretty good, once a second feels too
slow. Note that we use the actual FPS data, not the cached data (which is
updated every 2 seconds only).