Redesigning BosonStarting::start(): The starting is now subdivided into "tasks". Every task is represented by an object of a class derived from BosonStartingTask.
Every task has a "taskDuration()" method which gives the estimated amount of
time that this task will take. The number is an abstract "time", i.e. there is
no real unit (second/minutes/...) assigned to it, it is just a number.
All these durations are sumed up for the progress bar and whenever a task is
completed, the current duration sum is reported to the progress bar.

This provides a very dynamic approach, hardcoded values all reside in the
BosonStartingTask classes now, which are comparable easy to maintain.

Additionally we support a "sub-task". Every task may report to the progressbar
that it enters/leaves a "sub-task". The text assigned to that sub-task is then
assigned as well (and of course the duration can be updated using sub-tasks)