fixing events for our needs: - Implementing UWidget::getVisibleWidgetsAt(), which returns a list of all   widgets visible at a given point. This is very important for layered panes! - Instead of delivering events to a the getVisibleWidgetAt() only, the events   are delivered to all widgets returned by getVisibleWidgetsAt() now.   Since every child may propagate an event back to the parent, the   dispatchEvent() method is "locked" once an event is received. Therefore after   dispatching is completed, dispatchEvent(NULL) has to be called to unlock   delivery.
This should fix chat and commandframe widgets mostly.
There is still one problem left: select a unit and click a cmdframe button. The
button click is not recognized and instead the unit below the cmdframe button is
selected (or no unit at all, if no units is below the button).
As a workaround: move/rotate/zoom the map so, that no part of the map is below
the button you want to click (i.e. only black is below it). Everything should
work correctly then.