Another implementation of event delivery: The previous scheme had one major disadvantage: it first delivered to the topmost widget, then (unless the event was consumed first) to the parent and _then_ to the other children, i.e. the siblings of the topmost widget.
This patch now builds up a stack of visible widgets using a DFS. The events are
then delivered to all widgets on the stack until one widget consumes the event.

Therefore (in the case of layered panes) an event is first delivered to the
top-most layer, then to the other layers, down to the bottom layer and finally
to the parent (i.e. the layered pane itself).

This fixes the mouse click problems with the commandframe.
One final problem that I am aware of is left: in the newgame widget, when you
open the combobox and click the "human" item, which is above the "Add AI player"
button, the button is clicked as well. However this is probably a problem of the
popupmenu (it probably doesn't consume the event correctly)