Add profiling
Note that these classes are very core and very central in boson and most of the
time people will just assume that most methods in here take "no time". So they
should be really fast. Only the actual process*() methods are usually expected
to be performance relevant, as they do the actual tasks.