Adding a profling graph widget. It looks much less nice than I had hoped, but maybe it'll be of some use anyway.
This graph displays the profiling entries of the past 2 seconds. Every entry is
a single line, every profiling type ("updageGL", "advance", ...) appears on a
different height, with a label describing this type below it.

The length of every line is calculated from the duration of the profiling entry:
a profiling entry that takes 2 seconds would be represented by a line covering
the whole screen width.