Delete any previous widget deleter in setBoUfoWidgetDeleter().
The widget deleters are supposed to delete the BoUfoWidget object when the
ufo::UWidget object is deleted. However sometimes it may be necessary to delete
the BoUfo object (not necessarily BoUfoWidget, as BoUfoAction is part of BoUfo
as well) without deleting the ufo::UWidget object first. In that case we need to
unset the widget deleter (otherwise the ufo::UWidget deleter causes a double
deletion).