Adding libbomath.
This library consists of all the classes from the previous bo3dtools.* and
bomath.* files, but they are split up in separate files.
In this library there is no use of Qt/KDE/Boson classes/methods allowed, even
bogl is disallowed here.
This library is supposed to be a general purpose library, i.e. one can write
small test programs with it without linking to libboson.la. Only qglobal.h is
included here, which is required for Q_INT32 and so on, but no Qt methods are
used. So it is sufficient if that file is present, no linking required.

All additional Qt/KDE/Boson dependencies should go to files in code/boson (such
as code/boson/bo3dtools.*)