Finally implement renderCamera(). This method draws a part of the currently visible view frustum onto the minimap (a part only because the part with z <0 is not relevant).
First we retrieve all 6 planes from the view frustum and calculate the
intersection lines (we have lots of them, but we require 4 only). Those lines
are then used to calculate the vertex points of the frustum, i.e. the corners.
This is done by calculating the intersection point of a line with the two planes
that the line goes through.

Then we intersect the resulting frustum with the z=0 plane, by that is we
move every corner so, that it is z >= 0.0

Finally we map the vertices (which are in world coordinates) to the minimap
coordinates and draw them onto the minimap.
Sounds easy, hm? Well, it was not ;-)