Make sure that we find the libGL/libGLU libraries if they are installed, even if the system does not provide a libGL.so/libGL.so file (debian does so by any weird reason - unless the -dev package has been installed).
We try to emulate the dlopen() search order (LD_LIBRARY_PATH, then
/etc/ld.so.cache (we use ld.so.conf instead), then /lib, then /usr/lib).
But instead of just loading the specified filename, we also try to guess the
correct filename, i.e. when we search for libGL.so, then we try all filenames
in that directory that match libGL.so*, i.e. including libGL.so.1, ...