Use sendEvent() instead of send*Event(). All send*Event() methods are protected now (except of sendEvent() of course).
In sendEvent(), we update the keyboard state before delivering a mouse, wheel or
key event. This makes sure that we always have the korrect settings when an
event is delivered, even if the korresponding key press/release events were not
delivered (e.g. because they were eaten by the window manager)