Special handling for upgrades of Speed, AccelerationSpeed and DecelerationSpeed. I did not write automatic upgrades of the "current value" of a property (upgrades always apply to MinValue or MaxValue only), therefore we need special code for every property of that we want to update the current value after an upgrde.
I consider these three properties a special case, as they are actually handled
and stored in BosonItem (which does not provide upgrade features). I hope
special code will not be necessary for any other upgrades.

Note that properties that use "factors" for their current value (such as e.g.
health) never need to update their value after an upgrade, because this always
happens automatically (as the actual value is calculated on the fly anyway)