return (0,0) instead of maxSize in getPreferredSize(), if maxSize.w or maxSize.h is 0.
this fixed the layout bug in the commandframe widget - when a previous child of
a HBoxWidget (same for VBox) took away all space in horizontal direction,
it'll be of size (0,0) instead of (0, maxHeight). (0, maxHeight) makes no sense
anyway, as it'll be invisible anyway due to width==0. all it would do is to take
height away from other widgets.