reverting the libufo cvs backport
this fixes left-mouse-button events when the game is paused.

the reason for this is that when the game is paused, the "game paused" label
takes the whole widget (most parts transparent). so the event is delivered to
that widget, which in turn delivers it to the focusmanager, which consumes it
(requesting the focus for the widget). so it is never delivered to any parent