Updating BosonItemRenderer animation mode design: BosonItems provides a virtual method getAnimationMode() that is supposed to return the animation mode of the item in its current state. BosonItem::updateItem() should be called whenever the animation mode might have changed. It uses getAnimationMode() to retrieve the current animation mode.
The item renderer updates the actual animation in BosonItemRenderer::animate().

With this scheme the animation mode is even available when the item renderer
does not make use of it (e.g. with --nomodels). Also it would be easy to set a
different item renderer on the fly while the game is running. But most
important: it reduces dependencies of BosonItem on the item renderer.

Additionally some methods of BosonItem got removed as they are fully provided by
BosonItemRenderer.