Moving the SpeciesData out of SpeciesTheme. From now on SpeciesData is meant to contain SpeciesTheme related data that is displayed somewhere or is in some other way relevant for the UI/GUI only (but not for the game engine).
For this, I introduce the class BosonViewData. Access to this class should be
made from within the gameview only (where commandframe and *renderer subdirs are
actually part of the gameview!)
This class provides access to a SpeciesData object for a given SpeciesTheme
pointer.

It also contains the BosonItemContainer objecs now.

The BosonViewData class is meant to contain all data (or rather objects of
container classes providing access to such data) that belong logically to
some object from the gameengine (SpeciesTheme, BosonItem/BosonShot/Unit, ...)
but which does not belong to the gameengine (models, textures, pixmaps, ...)