splitting up BosonGroundTheme:
BosonGroundTheme now contains the BosonGroundType containers (as it did before).
These contain all values from the config file, but no actual data objects (they
did before).

BosonGroundThemeData always belongs to exactly one BosonGroundTheme. It contains
BosonGroundTypeData objects, which are similar to BosonGroundType: for every
groundtype, there is one data object. These objects provide the actual data
objects (textures, shaders, icons, ...)

The BosonGroundThemeData is created and owned by BosonViewData and is nearly
similar to the SpeciesData objects, except that the BosonViewData actually takes
ownership of the groundtheme data objects.
(we might one day do the same for SpeciesData, but atm I don't see the need for
changing working code)

With this patch the gameengine is completely separated from the GUI, that is we
follow a real document-view model now.
This is demonstrated by bounit, which links to the gameengine, but does not link
to libboson.la anymore.