Consume MouseDragged events in pushbuttons.
MouseDragged events are sent to the button when it is pressed down and the mouse
is moved (before the button is released again).
This fixes selectionrect problems when clicking a button: when the mouse is
moved a bit while clicking the button, the event is propagated to the parent of
the button which draws a selectionrect (which never disappears, as the button
eats the button release event).

This should fix bug #107646.


CC'ing libufo-devel, as this problem probably appears in official libufo, too.
However we use our own event code which also works with layered panes, and I did
not want to write yet another test application just to test whether the problem
appears with official cvs, therefore I am not patching libufo directly.

CCMAIL: libufo-devel@lists.sourceforge.net