in quitGame(), remove the effects from the itemcontainers before deleting them.
quitGame() is called on the widgets _before_ the actual game quits, i.e. before
the items are removed. Therefore the effects are deleted before the items that
use them are. This is contrary to the way it works in normal game, as an effect
is never deleted before the items stop using it.
However this scheme is necessary, as calling quitGame() on the game before
calling it on the widgets would imply that no cleanup of players/items/canvas
could be done anymore, as these would already be delted. Therefore a quitGame()
of the widgets would become useless and destruction code would become much more
tricky.