provide the ufo::UBoLabel as parameter to the base class, instead of adding it in init(). This makes sure that e.g. setKeyEvents(false) works on the actual label, instead of the container widget the label is in. This fixes shortcuts in the game: previously when pausing the game and then clicking in the game view, the shortcuts did not work anymore (as the game thinks the label actually wants the key events).
This is a sub-optimal solution only. The shortcuts will probably cause problems
again later.
Ideally we would have a solution that is comparable to the one in Qt (that I
once implemented in libufo, but never was accepted to upstream) that uses
AccelOverride and Accel events before sending the actual key events.
This way we could make lineedits/textedits accept AccelOverride events and be
happy. Other widgets that need key events would still receive them, but
shortcuts (Accels) would have precedence