testing map, showing relations between different models
This shows that the size of some untis and nearly all non-neutral buildings
needs a change.

Of course, it would also be possible to make the units smaller and leave the
buildings. IMHO (and Rivos) the neutral factories are correct. Tanks fit in
there also. Things that need a change are e.g. wolf, toilet