Changed the size of many models because they don't fit together.
This makes most maps unusable and might cause other problems.


I am not sure if it is good to commit such a change before a release, but
since new maps are on the way this should be done before maps get created.

I am also not sure if this doesn't need some discussion first, so feel free
to see this a request for comments.


You can see the change on the testing/relations map. Get the map from cvs
and look at it. Than update themes and looks again on it. I think it is
better.


Btw. if someone can make a screenshot of the new relations map before
updating themes we can see the change. I missed that...