Implementing the new player IDs scheme, as described in PlayerIds in the wiki.
We use KPlayer::userId() instead of the normal id() now. This means we have full
control over the ID and we do not care anymore whether a player comes from
client A, client B or whatever. Also we can easily assign the ID on startup, so
that e.g. a map can provide 3 possible players and there can be 2 players
playing the map who can freely select which side they want to play on. The
remaining player is simply not played by anyone.
Previously that was possible too, but much harder to achieve.

Changes in detail:
- Make KGame::systemAddPlayer()/systemRemovePlayer() virtual
- Rename KPlayer::id() to KPlayer::kgameId() and KPlayer::setId() to
  KPlayer::setKGameId()
- Rename KGame::findPlayer() to KGame::findPlayerByKGameId()
- Add KGame::findPlayerByUserId()
- Overwrite KGame::systemAddPlayer() in Boson.
  The new implementation first sets the (user) ID of the player, if not set yet,
  and then calls KGame::systemAddPlayer().
- Adding a file converter that does the job of BosonStarting::fixPlayerIds(): it
  converts the player indices (that are stored instead of the IDs in .bpf/.bsg
  files) to actual (user) IDs.
- Adding Player::bosonId() that returns KPlayer::userId().
  This is the preferred way of accessing the ID of a player currently.
  Boson should NOT use KPlayer::kgameId() directly. The KGame player ID is
  required for sending messages only.