adding visual feedbacks: whenever the (local) player orders units to move to some point, a small green cross is rendered at that position for about half a second, to indicate that his order has been understood. whenever a unit is ordered to attack another unit, the attackes unit is tinted red, indicating that this unit will be attacked.
adding further feedbacks is very easy: add a signal to the localplayerinput that
is emitted when the input occurs, connect it to a corresponding slot in the
canvasrenderer and create a feedback object there. thats it. it can be done
completely by doing copy&paste and changing the method names.

adding new feedback types is kinda easy, too: derive from BoVisualFeedback and
reimplement rtti() and paintGL() (or integrate it somewhere in the rendering
code, if an independent paintGL() is not applicable()).