adding a quicksave (F5) / quickload (F7) feature
note that the quickload takes several seconds and it is unlikely to get any
faster somewhen soon. most time is spent in XML processing and in initializing
the pathfinder, both are mandatory tasks. in theory the XML processing could get
eliminated by using a binary file format - but... been there, done that.