Use setShields(shields() + 2) instead of setShields(shiels() + 1) for reloading. This fixes shields reloading.
The problem here is that we use a shieldsFactor for storing the actual shields,
which causes precision trouble: bofixed cannot store all values and so it is
possible that shields() and shields() + 1 are represented by the same
shieldsFactor due to rounding errors. This leads to setShields(shields() + 1)
being a noop in many cases and so we never reload anything.
The problem does not exist when using shields() + 2, as a rounding error of > 1
is not possible (note that "rounding" can also mean integer conversion, i.e.
0.999999 is "rounded" to 0).

The problem might also vanish once we move to floats again, but I am not sure
about this. But it would definitely happen _much_ less often.