oh i hate such days... i spent several hours developing a solution to this problem, only to find out that we already had one - it just didn't get used properly. and even worse: the best solution is not to develop new code, but rather to remove obsolete code
so here it comes:
removing createVisibleCellList(). this method was used to filter out all fogged
cells from the renderCells(), so that only actually visible cells would get
rendered.

this was perfectly working when it was written, but had problems with recent
ground renderers: instead of rendering a single cell, often a rect that covers
several cells is rendered (usually due to LOD).
with that createVisibleCellList(), either displays the whole rect, or nothing.
fixing it would require some major design changes, as the groundrenderers
(especially the default one) currently do some huge assumptions on the cell
lists they generate and those that they eventually render. especially concerning
heightmaps.

but fixing it is not requied anyway. the fogtexture approach that rivo added a
(pretty long) while ago implements a much better fow anyway. by not removing the
cells from the list, we don't show the player anything that he shouldn't see, as
the cells won't ever appear on the screen.
the only problem is, that these cells are always rendered, even if they will
eventually be overdrawn by the fog texture. but i believe that this is a very
minor issue - in a usual game there aren't so many fogged areas visible that
they could become a problem.

even if they do - the best solution would be to include the fog of war directly
into cell list generation, not as a separate method afterwards.