implementing the ammunition scheme as discussed on irc: * every weapon has a "RequiredGenericAmmunition" field that specifies how much   of the "generic" ammunition is required per shot. if the player has no   ammunition left - the unit won't be able to reload. * the "generic" ammunition is delivered to the units automatically, they don't   have to pick it up * the ammunition is taken from a "GenericAmmunition" pool that every player has * every advance call the player receives new (generic) ammunition, as long as   he has < 1000 ammunition.
* the rate in which the player gets new ammunition could be increased by
  building an "ammunition center", however this is not yet exisiting.

for now all weapons use the same type of ammo: "GenericAmmunition". i decided
not to implement a we-support-as-many-types-as-we-want scheme, as such things
often turn out to be overly complex and don't fit our needs.
instead i imagine that we add new types by duplicating code for now: next to
"mGenericAmmunition" we add a "mMissileAmmunition" or so. once it turned out
what we actually need, we'd then refactor the code to make it nice.