First small step towards turret support. Every weapon can have a "TurretMesh" key that specifies a mesh that is used as turret. A turret is supposed to be rotated into the direction where the weapon shoots to.
This "TurretMesh" key is represented by a BosonWeaponTurret object of a
BosonWeapon. It can modify a rotation matrix, that is applied to the
corresponding mesh before rendering it.
borender demonstrates the rotation of such a mesh currently - it rotates a mesh
named "turret" every frame.

Note that this is not yet the final implementation - it probably needs to change
drastically.
The problem is that boson does not support hierarchical meshes, which are
required for this feature. This can be seen in borender currently with the
"turret" model. That model has a "turret" mesh, that is meant to be rotated.
However the "gun" mesh, that should be rotated along with the "turret", is not
rotated - it remains at its original position.
I don't see how we could fix this problem with the current BosonModel design, I
am pretty sure we need hierarchical meshes :-(