I actuall found a way to implement turrets without implementing hierarchical meshes (aka rewriting large parts of the model code).
The index.unit files have a "TurretMeshes" key now, that that takes a list of
meshes that are to be rotated. The models are supposed to use hierarchical
meshes (in the .3ds file) - this entry should contain all of the meshes in the
hierarchy that should be rotated. Boson will then behave as if hierarchical
meshes would be used, the rotation applies to all of these meshes.

borender already demonstrates this correctly for the "turret" model.

However note, that in boson rotations are not yet working, as
a) the code is completely untested (I cannot start boson currently, due to
   problems with my system)
b) the actual rotation is not yet made. will be impleemnted soon.