adding a dialog that aids in creating TurretMeshes entries
it parses the .3ds file to present a listview with all meshes. there the models
that belong to the turret can be selected and a TurretMeshes line is generated
(and can be edited manually).

it is integrated into borender - using the apply button the changed turret can
be tested immediately.