Interesting what I can find on my harddisk ;-)
I wrote this class a long time ago for a different purpose that was not followed
further. Now it is of some use again, so here it comes.
The pixmap renderer takes a GL widget and applies a different context to it,
that is based on a QPixmap. Therefore it is similar to renderPixmap(), however
it allows to render several frames, not just one and therefore is much more
efficient if more than one pixmap should be created.