adding BoModelPixmaps.
This program is primarily meant to render a model to pixmaps. It is used to
create a gallery of the model for APALOM, see
http://www.freehackers.org/~andi/apalom/

It also collects all textures that the model uses and checks for the copyright
of all models and textures. Finally it allows to create a "package" with all
these files.

Such a .tar.gz archive could easily be downloaded and used in any (open source)
game.