attempt at fixing player ID assignment for network games.
when a client connects, the ADMIN now first adds the new players (with the ID
that was provided by the client, but is invalid in the network game), then sets
the ID to 0 and finally calculates a new ID for that player.
after that the normal network connection procedure of KGame proceeds and ensures
that the new IDs (provided by the ADMIN) are used at client side, too.

this patch is untested, the above describes how it _should_ work.