support for the .md2 format.
.md2 is a very simple format, that supports one texture per model and has basic
support for animations. However it does not use a frame matrix like .3ds does:
every frame needs to provide a complete new object.

md2 has a lot of limits that makes it barely useful for modern high quality
models, however it is sufficient for "good" models and there are many models out
there in this format.
In addition this restrictions help at making games fast ;)