adding a copy of the bksys directory from kdelibs in KDE trunk.
the bksys directory is an equivalent to the "admin" directory in the bksys build
system (which is based on scons).

a few files are slightly modified (e.g. kde3.py has .boui support) but pretty
much all of my additions for boson are in separate files, such as boson_*,
openal.py, libufo.py, ...