get rid of the libboson.a library.
this has been a helper library only, without any real use. we don't need it at
all with scons.

this also removes a couple of scons warnings of building the same file in
different environments. however this is not actually a fix, but rather a lucky
coincidence: I already wrote a workaround (more precisely: an evil hack) for
these warnings - the workaround just didn't work anymore when building the file
in both, a static library and a program.