adding libcommon_libs and libgameengine_libs that list the libraries that libcommon.a/libgameengine.a depend on. previously libboson_lib contained all of them in a single big list, but that is very unpractical.
in theory it's now possible to build boson-no-gui without linking to most libs
(even though bogl and imageio are still necessary. but that's a dependecy of
boapplication.cpp)
however in reality we still link against pretty much all libraries, because of
the "kobject.env = env" hack, that removes _many_ build system warnings. a clean
fix for that problem (instead of copying the environment) is yet to be found.