dlopen() sucks. don't let it do magic - we are better magicians. i hope.
dlopen() expects *.so filenames that don't exist on all systems, often we have
*.so.1 only. therefore we already search for such files, if dlopen("GL") fails.

however the problem is getting _much_ worse when by any reason the correct
library is not found (e.g. *.so.1 filename), but a different library is found
(e.g. a concurrent libGL.so that exists on the system, but does not belong to
the correct driver). then dlopen("GL") returns no error and so we won't try our
workaround (which might probably work).

so instead of letting dlopen() decide what to load, we always search for the
correct path on our own and give an absolute filename to dlopen() only.

in addition we now support "include foo/*.conf" entries in /etc/ld.so.conf that
fedora core uses.