fixed the most important outstanding issue: the "different environments for same target" problem. it turned out that SCons allows to suppress warnings - so we just do that for this warning.
of course this is not a "clean" fix, but it is one without evil side effects. if
the same target should be built using different commands in the environments,
then SCons will emit an error, not just a warning, so we should be safe.

this especially removes the side effect that lots of unrelated libraries are
linked in binaries (as we kinda re-used environments).

as a result i had to fix some library dependencies - more or less a little bit
of cleanup.