adding LOD::removeAllMeshesBut() and removing LOD::setMeshes().
an LOD object is now responsible for all of its meshes and no other object needs
to (or is even allowed to) mess with that. for this adding a Frame::removeNode()
method that makes sure a frame object does not reference the mesh anymore.

this removes some duplicated code around the setMeshes() calls and in addition
it makes sure that there is no dangling pointer in a frame to one of the old
meshes.
in particular it fixes a TODO on this topic in the mesh optimizer