Rewriting the IdStatus messages: Previously they sent out _every_ game event. That would be one per advance call at an absolute minimum (!) and no maximum number restriction. That means we had a network traffic consumption linear to both, time and game activity. And this for updating the unit counts of the dedicated server only.
Now we maintain a statistics class that sends out a status message once in a
while containing the relevant information (atm unit counts only).
This class can easily be extended with more information, but I think it might be
easier to simply have a game instance running at the server, that the server may
use to retrieve the information.