Adding BoCanvasQuadTreeNode.
This tree provides (in addition to BoGroundQuadTreeNode):
- unitMinZ()/unitMaxZ()
  these methods propvide the minimal z value (z()) and the maximal z value (z()
  + depth()) of all units in this node
- minZ()/maxZ() - the minimum (maximum) of unitMinZ() and groundMinZ()
  (unitMaxZ() and groundMaxZ())

The values are updated in BosonCanvas::removeFromCells() and
BosonCanvas::addToCells() which ensures they are always correct. However this is
also pretty inefficient: in many cases removeFromCells() is called, followed by
addToCells(), without anything changing. Or these methods are called for many
units and we update the values for every single change.
A more efficient implementation would store a "dirty" flag and calculate the
values whenever minZ()/maxZ()/unitMinZ()/unitMaxZ() are called.
But for now it seems to be quite ok: when moving units, about 6% of the time are
spent in the tree, which is a rather small value only I think - at least as long
as we have only a single tree.
Therefore I leave it at this for now.

Note that in CVS code no such tree is currently being created.