New coordinate mapping architecture. Previously we used glReadPixels() retrieve the depth at the mouse position. From that we calculated the z coordinate and then we had enough information to calculate the (x,y,z) values. This approach had several disadvantages, such as * some drivers are broken when reading depth values (hello ATI) * it is slow. A glReadPixels() call took around 1ms which is _much_ too slow for   using it in mousemove event handling (you could notice that CPU usage went up   significantly when moving the mouse) * the selection rect is _very_ tricky: in theory we would have to read the depth   for every pixel inside the rect, calculate the position and find all units at   these positions. That would be terribly slow.   That is why the selection rect has often been very imprecise - especially when   trying to select flying units with a rotated camera.
The new approach does not use OpenGL at all anymore. Instead we emulate the
"picking" behaviour (i.e. GL_SELECT mode) of OpenGL:
1. We calculate a picking matrix from a given widget-coordinates rect (or point)
2. This matrix is pre-multiplied with the current projection matrix
   If that new matrix was used as projection matrix, it would cause the whole
   screen (i.e. the viewport) to display only the area that was specified by the
   rect - i.e. it zooms/translates so that the selectionrect fills the viewport.
3. With this new projection matrix we construct a viewfrustum
4. That viewfrustum is used to find the items inside the given rect.

The last step is the crucial one - if this is not _very_ fast, then the whole
algorithm will be slow. Fortunately we do that in every frame anyway (to
retrieve the list of visible items), so it must be very fast anyway (and it is).

However this will give items only, not the ground position (this is required
only, if we search for a single point, not a rect)
For this we use a similar separate algorithm. Instead of step 4 above we use the
viewfrustum and a ground quadtree to find the visible cells. These are then
subdivided into another (virtual) quadtree a couple of times until the
precision of the coordinates is high enough (about 1/8 cell atm).



This new code is very experimental but seems to work so far.
Currently click-selection of a (single) unit is broken however. The new code
calculates the _ground_ position under the mouse - if the camera is rotated,
then it is unlikely that the unit under the mouse is at that position.

This is however only a design issue, it will be fixed soon - we simply should
not use the calculated position to find the unit under the mouse (but rather the
new picking architecture)